chore: create component dir for sidebar and move files
